xci: Bump Bifrost components to the HEAD of Rocky release
The bifrost and ironic SHAs are from the Stein release which is still
in early days of development. We should use Rocky SHAs for the time
being. However, bifrost from Rocky does not support Ansible > 2.4 so
we have to use bifrost from master branch but all the ironic components
have to be from the Rocky release.

Change-Id: I1acb28e4a292c774d681519517be19a2da230851
Signed-off-by: Markos Chandras <mchandras@suse.de>