CN should point to vip address
When ssl handshake happens, the client will verify the server
certificate and will try to match the Common Name (CN) of certificate
with the domain name in the url, if both are different, hostname
verification will fail. This is what was currently happening and CN
should point to the ip (or domain name if we had DNS) or the
external_lb_vip_address. Unfortunately, at this point the variable is
not available, so I guess we need to hardcode it

Change-Id: I81c5ba78148a9286d03a84db5cd86a0bac46507f
Signed-off-by: Manuel Buil <mbuil@suse.com>