Unfreeze networking-odl and networking-bgpvpn
There was a problem with ceilometer which was the reason why these two
repos were frozen. Unfortunately, freezing networking-odl breaks the
integration between neutron and ODL in rocky. Fortunately, the problem
with ceilometer dependency is now gone, so as far as I see, we can
unfreeze these two repos. The patch to unfreeze was already created
upstream but as everyone is at the PTG and gates are broken, it might
take a while until it gets merged:

https://review.openstack.org/#/c/602243/

Without this patch, all scenarios working with ODL will fail when doing
the healthcheck verification with the error:

Unexpected API Error. <class 'neutronclient.common.exceptions.ServiceUnavailable'>", "code": 500}

Change-Id: I05e7f28b5c2a28baa632b760436ffbff17e93b18
Signed-off-by: Manuel Buil <mbuil@suse.com>