Fix the ODL scenarios for Rocky
Since Rocky, networking-odl depends on ceilometer and requires it to be
installed. Therefore, all odl scenarios need to have ceilometer
deployed. Once that is done, we can unfreeze n-odl repo.

Besides, we need to introduce a SHA bump for neutron and ceilometer to
include the latest changes to support this fix

deploy-scenario:os-odl-nofeature
installer-type:osa

Change-Id: I81a39436e4ff648faabda4e82fce1d3f14615741
Signed-off-by: Manuel Buil <mbuil@suse.com>