Simplify external networking for os
We were right now creating an overlay network with gateway in the opnfv
vm. This makes floating ips work but if we need internet connection from
vms, things get complicated and messy.

As we are only using up to 7 ips from the 192.168.122.0/24, we can
consider it the provider network for openstack and use ips starting from
192.168.122.100 as floating ips or ips for routing (doing SNAT as
always)

Change-Id: I09af663069ae95a9d265d98f1531778eb37134e2
Signed-off-by: Manuel Buil <mbuil@suse.com>