[baremetal] Provide networking for suse nodes
When deploying baremetal, the traffic should be segmented using the
different interfaces instead of through vlans. All the config is done
based on idf and pdf information.

When doing non-baremetal, opnfv and nodes get the same config.
When doing baremetal, opnfv and nodes get a different network config

Apart from that, if vlan_id is defined in the name, there is no need
for VLAN_ID in the interface descriptor. This simplifies things

Change-Id: Iddbb90af807b43e247e5ee11fe735df9e823d4bf
Signed-off-by: Manuel Buil <mbuil@suse.com>