Bug fix: Fetch vm names from idf
Currently we were taking the names of the nodes to be generated from the
NODE_NAMES variable whereas ansible was fetching the names from the
dynamic_inventory.py which uses idf. This resulted in problems: when
doing ha, ansible was provisioning a compute as a controller and
vicebersa.

This patch forces create_nodes role to fetch the name from idf and thus
align with the naming schema of ansible

Change-Id: Id1473727405701fd9ed0cb2f1394ee8676cec337
Signed-off-by: Manuel Buil <mbuil@suse.com>