Simplify the serial config
The current configuration only dumps the booting info into a file and
the pty does not work (i.e. virsh console opnfv returns a failure
because it cannot find a character device). After some investigations,
it is apparently impossible to have both active:

https://github.com/Mirantis/virtlet/issues/249

Therefore, we should remove the pty part of the xml. To connect to the
VM in case of network problems, we can always use vnc.

Apart from that, the console part is not necessary as libvirt will
create that one for us

Change-Id: I80a59163b4ba4e6bff34cb5378893201e93ddb87
Signed-off-by: Manuel Buil <mbuil@suse.com>