[WIP] Introduction of Openstack-helm as installer
This patch creates a new installer tree in xci that uses
openstack-helm to deploy openstack on a kubernetes cluster.

This is Work-in-Progress and mostly created to facilitate
discussion, while development is ongoing. Documentation
also not created yet.

USAGE: Export INSTALLER_TYPE=osh, DEPLOY_SCENARIO=k8-calico-nofeature
and XCI_FLAVOR=noha and run deploy.sh as in documentation.

STATUS: Works end-to-end. Builds demo stacks and validates with tempest.

Change-Id: I212f70eb51c2a38c798c11367d2ebb8bf5f4a1de
Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>