Update patch set 1
Patch Set 1:

(5 comments)

thanks for the patch Nikos.

as you will notice, I only commented on opnfv-scenario-requirements.yml as what we are doing requires changes in framework itself first to adapt it to handle multiple installers that deploy OpenStack scenarios or Kubernetes scenarios.

Apart from that, the scenarios deployed by OSH should be named accordingly. Yes, we should (re)use the existing k8s scenarios deployed by Kubespray but we can do that perhaps in another way like creating a new os-nosdn-nofeature scenario for OSH, taking whatever comes from k8s-nosdn-nofeature and adding OSH steps on top of that scenario as additional tasks.

i would like to hear Markos and Manuel as well and what I am thinking might not be the correct and/or best way to handle this new situation.

Patch-set: 1