Fix bug in the inventory generator
The dns keys should only appear in case the idf provides a dns entry for
the network. Otherwise, it should not exist at all. If it exists, even
if the value is empty, ansible will transform it in a variable and

"item.network.dns is defined"

will return true:

https://github.com/opnfv/releng-xci/blob/master/xci/playbooks/roles/bootstrap-host/templates/osa/debian.interface.j2#L35

A docstring is added to explain what is the purpose of the class

Signed-off-by: Manuel Buil <mbuil@suse.com>
Change-Id: Ib8afa06cecb54f384083060073fa463c7f8d313f