Delete keepalived folder for OSH/Armada xci-type
In case that the XCI is deployed with type armada or osh,
the 'keepalived' folder is created. Before each redeploy of XCI in the
same server, the folder should be deleted in order to be created again
successfully.

In case that the XCI deployed with other type than armada or osh,
this folder is not used and there is nothing to be deleted.

Change-Id: I698b8125464a93faca3b41ad75db9ec80696883b
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>