Update patch set 14
Patch Set 14: Code-Review-1

(9 comments)

WAW! This is a big and difficult patch but in general looks good, congratulations! Sorry for the delayed review but I needed to understand OSH first.

First of all, could we split the patch into at least two (one would depend on the other)? For example, one that handles the OSH deployment and other which takes care of the idf, create-nodes role changes that are required to deploy k8s? That would certainly help the reviewers and reduce the complexity.

Second of all, I'd use an ansible role to implement the install-openstack-helm instead of a playbook, it will be much easier and scalable for future things. In fact, I think we could merge both install-openstack-helm-mini and noha in a not very difficult way by using a role (using the inventory, using templates and jinja2...)

Third, in that role, we should use ${XCI_PATH}/xci/playbooks/dynamic_inventory.py and use the names of the nodes instead of static ips. We could for example group all computes into a group called "computes" and there we can execute all tasks that target computes instead of having to loop over ips and use "delegate_to"

Last thing, according to my investigations, if we deploy OSH multi-node, there are some files which we must change, for example, in values.yaml of neutron we should define the tunnel interface for the vxlan tunnels to work.

Patch-set: 14
Label: Code-Review=-1