Update patch set 14
Patch Set 14:

(9 comments)

Thanks a lot for your valuable comments.

On the first three of the general comments, here, I agree and I'm working on restructuring the patch according to these principles. Just to say that I wanted to have a known and successful path of actions first and then reshape it in a proper way, rather than troubleshooting both in ansible and in OSH at the same time.
On the last point, I'm not sure on how far this patch should go: the general idea that I followed is that we're taking the OSH, as unchanged as possible, and putting it on XCI, this including the validation steps as provided by the OSH. I noticed that networking in multinode is not in perfect shape, but shouldn't we try to fix it there if possible?
Thanks again for your time and effort for this input.

Patch-set: 14