Update patch set 22
Patch Set 22: Code-Review+2

I think it is good to be merged! There are things we must improve but this is a good foundation to start work. Let me try to fix opensuse problems so we can get +1 from Jenkins (anyone wants to help?)

List of things to improve (Nikos, probably you can take these as next steps):

* Find a way to deploy scenarios (right now scenario must be k8-calico-nofeature)
* The last script we execute: "execute 900-use-it.sh", does not remove what it creates. We should either go upstream and change that or move that script to the testing part
* Add a guide to explain how to operate this (what you explain in the commit message + export OS_CLOUD=openstack_helm
* Any other you can come up with?

Patch-set: 22