[DNM] Move to python3 in the deployment host
JIRA:XCI-2

deploy-scenario:k8-calico-nofeature
installer-type:osh
xci_distro:ubuntu_python3

STATUS
======
It is not clear how distros will react to py2 expiration. Some will
convert the package names so that python- defaults to python3-, some
will not. Having said so:

* Bifrost packages are still python-xxx and not python3-xxx.
https://github.com/openstack/bifrost/blob/master/playbooks/roles/bifrost-ironic-install/defaults/required_defaults_Ubuntu_16.04.yml
https://github.com/openstack/bifrost/blob/master/playbooks/roles/bifrost-ironic-install/defaults/required_defaults_Ubuntu_18.04.yml
https://github.com/openstack/bifrost/blob/master/playbooks/roles/bifrost-ironic-install/defaults/required_defaults_openSUSE_Leap_15.0.yml

We need to wait and understand how the distros will react to understand
what needs to be done (maybe a patch in Bifrost is needed). Besides some
of those packages (e.g. python-configparser) have a dependency with
python2 and python2 gets installed.

* pip2 is default right now. When using python3 for the
dynamic_inventory.py, yaml is missing (jenkins error) because it is
installed in pip2 and not in pip3.
To fix that, use XCI_DISTRO=ubuntu_python3 to use an image that has pip3
as default. Let's wait to see how distros react to this and see if they
use pip3 and python3 as default.

* openSUSE Leap 42.3 is not supported anymore and it uses Python3.4 for
python3 support. Ansible requests Python 3.5 or greater. So we'd better
try to provide support for Leap 15 and then move to Python3

Change-Id: Iabb9345bb3716156ace2b6e3814e13fa8aa9dadc
Signed-off-by: Manuel Buil <mbuil@suse.com>