Remove the need to specify scenario when deploying openstack-helm
The XCI openstack-helm installer is supported only for a specific kubernetes
cluster network scenario (k8-calico-nofeature) and having to specify it serves
no purpose. This patch removes the need to supply DEPLOY_SCENARIO when deploying
with OSH and enforces the only available option.
Moreover, in the futute, the DEPLOY_SCENARIO variable is able to be used as intented,
to specify a scenario for the openstack deployment on the kubernetes cluster.

installer-type:osh

Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>
Change-Id: I238e02bd18daeb3b5222c5582a45e461eda550f2