Refactor k8s scenario implementation
This change updates the scenario logic for basic k8s
scenarios that enables the use of various network plugins.

The change simply stops copying k8s-cluster.yml over but instead
operates on the existing/original k8s-cluster.yml and updates
kube_network_plugin in place.

This makes it easier to maintain the scenarios and more importantly
unblocks us from bumping kubespray and kubernetes versions.

Please note that opnfv-scenario-requirements.yml contains version and
the refspec of the corresponding change in releng-xci-scenarios repo
so this change must be verified properly and if it passes CI, it
needs to be amended, updating k8-calico-nofeature scenario's version
to master and submitted. And then the change in releng-xci-scenarios
can properly be verified.

https://gerrit.opnfv.org/gerrit/#/c/releng-xci-scenarios/+/68418/

installer-type:kubespray
deploy-scenario:k8-calico-nofeature

JIRA: XCI-20

Change-Id: I1f851d20f5f7bc256199303581b05659091fddf9
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>