Enable the use of multiple installers per scenario
A new installer, OpenStack-Helm is being introduced to XCI and this
requires the XCI Framework to have capability to use the right/specified
installer for given scenario as multiple installers could support a single
scenario. An example to this is os-nosdn-nofeature scenario which is supported
by both OSA and OSH

This change simplifies the XCI scenario collection and plugin logic by
cloning releng-xci-scenarios repository once, checks out the version
specified by the scenario in opnfv-scenario-requirements.yml file as
we don't support concurrent deployment of multiple scenarios.

The corresponding change adapts the directory structure is sent for
releng-xci-scenarios repository.

https://gerrit.opnfv.org/gerrit/#/c/releng-xci-scenarios/+/68432/

installer-type:kubespray
deploy-scenario:k8-calico-nofeature

JIRA: XCI-23

Change-Id: Ib3cd5219bc097b80ef132c51a1a91f04d94d7676
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>